A 5ms timeout seems a bit stupid?
I apparently skipped the 'read the manpage' step on that one...

Note that since we only update the changes counter on POLLERR events,
and not on every return (f.g., on timeouts), that timeout is pretty much
useless anyway, and we could just block (-1).

We don't have any practical use for keeping track of that timeout,
because we do want to block until the mountpoint is up, especially when
Nickel unmounts it for USBMS.
Right now, if we were to track that timeout, that'd meant giving up
after 7.5s, which is stupid in mostly all instances of our use-cases ;p.