Try to reap zombie processes *again*...
Before handling the next batch of inotify events.
Do so for the sole reason that the pipe dance doesn't appear to take for
every script out there on my device, for some mysterious reason...

Why it works perfectly fine on my desktop sandbox, and even on my device
with some scripts but not others, is beyond me.

That means we may hand on to *one* zombie per watch for a bit (until the
next batch of inotify events).
It's not pretty, but it doesn't hurt much (besides my pride).