Expand on that last comment tweak a bit...
Depending on timing, we may still say hello before on-animator's
first pickel call, in which case, the fb rotation is slightly different.
(i.e., on my H2O, that leads to printing in the right place,
but upside down).
I knew that FBInk release name was more than just a cute joke ;D.